Refactor naming to make code more readable - hotel overview
