+++ src/DOMelements.js
@@ -1,37 +1,52 @@
-export let backToBooking = document.querySelector('.back-to-booking');
-export let bookARoom = document.querySelector('.book-a-room');
 export let bookRoomButton = document.querySelector('.bookroom-button');
-export let bookingModal = document.querySelector('#add-booking-modal');
-export let bookRoomDate = document.querySelector('.book-room-date');
-export let bookRoomHeader = document.querySelector('.book-room-header');
-export let customerRooms = document.querySelector('.display-customer-rooms');
-export let customerStatus = document.querySelector('.customer-status');
-export let customerView = document.querySelector('.customer-view');
-export let customerWelcome = document.querySelector('.customer-welcome');
+export let customerDirectoryButton = document.querySelector('.searchcustomers-button');
 export let enterCredentials = document.querySelector('.enter-credentials');
 export let errorMessage = document.querySelector('.error-message');
-export let filterCategories = document.getElementById('filter-categories');
-export let filterRefreshButton = document.querySelector('.filter-status');
-export let filterSection = document.querySelector('.filter-rooms');
-export let filterSubmitButton = document.querySelector('.filter-button');
-export let hotelOverviewDate = document.querySelector('.hotel-overview-date');
-export let managerView = document.querySelector('.manager-view');
-export let modalContent = document.querySelector('#modal-content');
-export let modalDate = document.querySelector('.modal-date');
-export let modalTitle = document.querySelector('.modal-title');
-export let overviewInfo = document.querySelector('#overview-info');
-export let passwordInput = document.querySelector('.password');
-export let searchCustomerButton = document.querySelector('#search-customer-button');
-export let searchCustomerInput = document.querySelector('.search-customer-name');
-export let searchTitle = document.querySelector('.search-title');
-export let signOutButton = document.querySelector('.signout-button');
+export let homeButton = document.querySelector('.home-button');
 export let loginButton = document.querySelector('.login-button');
+export let managerView = document.querySelector('.manager-view');
+export let hotelOverviewInfo = document.querySelector('#hotel-overview-info');
+export let hotelOverviewTitle = document.querySelector('.hotel-overview-title');
+export let passwordInput = document.querySelector('.password');
+export let signOutButton = document.querySelector('.signout-button');
+export let todayHotelOverview = document.querySelector('.today-hotel-overview');
 export let usernameInput = document.querySelector('.username');
-export let viewBookingInfo = document.querySelector('.view-information');
-export let overviewTitle = document.querySelector('.overview-title');
-export let todayOverview = document.querySelector('.today-overview');
-export let todayBookings = document.querySelector('.current-bookings');
-export let gridContainer = document.querySelector('#customer-bookings');
-export let bookingsDate = document.querySelector('.bookings-date');
-export let customerDirectoryButton = document.querySelector('.searchcustomers-button');
-export let customerSearchBar = document.querySelector('.customer-search-bar');
+
+
+
+// export let backToBooking = document.querySelector('.back-to-booking');
+// export let bookARoom = document.querySelector('.book-a-room');
+
+// export let bookingModal = document.querySelector('#add-booking-modal');
+// export let bookRoomDate = document.querySelector('.book-room-date');
+// export let bookRoomHeader = document.querySelector('.book-room-header');
+// export let customerRooms = document.querySelector('.display-customer-rooms');
+// export let customerStatus = document.querySelector('.customer-status');
+// export let customerView = document.querySelector('.customer-view');
+// export let customerWelcome = document.querySelector('.customer-welcome');
+
+
+// export let filterCategories = document.getElementById('filter-categories');
+// export let filterRefreshButton = document.querySelector('.filter-status');
+// export let filterSection = document.querySelector('.filter-rooms');
+// export let filterSubmitButton = document.querySelector('.filter-button');
+// export let hotelOverviewDate = document.querySelector('.hotel-overview-date');
+
+// export let modalContent = document.querySelector('#modal-content');
+// export let modalDate = document.querySelector('.modal-date');
+// export let modalTitle = document.querySelector('.modal-title');
+
+
+// export let searchCustomerButton = document.querySelector('#search-customer-button');
+// export let searchCustomerInput = document.querySelector('.search-customer-name');
+// export let searchTitle = document.querySelector('.search-title');
+
+
+// export let viewBookingInfo = document.querySelector('.view-information');
+
+
+// export let todayBookings = document.querySelector('.current-bookings');
+// export let gridContainer = document.querySelector('#customer-bookings');
+// export let bookingsDate = document.querySelector('.bookings-date');
+
+// export let customerSearchBar = document.querySelector('.customer-search-bar');
