+++ src/css/base.scss
@@ -94,7 +94,7 @@ header {
 
 // Manager View
 
-.today-overview {
+.today-hotel-overview {
   display: flex;
   flex-direction: column;
   justify-content: center;
@@ -107,7 +107,7 @@ header {
   border-radius: 10px;
 }
 
-#overview-info {
+#hotel-overview-info {
   display: flex;
   align-items: center;
   flex-direction: space-around;
